........
  r323672 | twilson | 2011-06-15 10:09:51 -0700 (Wed, 15 Jun 2011) | 5 lines
  
  Cast ARRAY_LEN to size_t for ast_logging
  
  32-bit and 64-bit machines return different types for ARRAY_LEN(), so cast
  it before using in a format string.
........


git-svn-id: https://origsvn.digium.com/svn/asterisk/trunk@323673 65c4cc65-6c06-0410-ace0-fbb531ad65f3

+++ tests/test_db.c
@@ -193,12 +193,12 @@ AST_TEST_DEFINE(gettree_deltree)
 	ast_db_freetree(dbes);
 
 	if ((num_deleted = ast_db_deltree(BASE, SUB2)) != ARRAY_LEN(inputs) / 2) {
-		ast_test_status_update(test, "Failed to deltree %s/%s, expected %lu deletions and got %d\n", BASE, SUB2, ARRAY_LEN(inputs) / 2, num_deleted);
+		ast_test_status_update(test, "Failed to deltree %s/%s, expected %zu deletions and got %d\n", BASE, SUB2, (size_t) ARRAY_LEN(inputs) / 2, num_deleted);
 		res = AST_TEST_FAIL;
 	}
 
 	if ((num_deleted = ast_db_deltree(BASE, NULL)) != ARRAY_LEN(inputs) / 2) {
-		ast_test_status_update(test, "Failed to deltree %s, expected %lu deletions and got %d\n", BASE, ARRAY_LEN(inputs) / 2, num_deleted);
+		ast_test_status_update(test, "Failed to deltree %s, expected %zu deletions and got %d\n", BASE, (size_t) ARRAY_LEN(inputs) / 2, num_deleted);
 		res = AST_TEST_FAIL;
 	}
